Modified some

+++ backend/import_export/zip_unzip.py
@@ -1,11 +1,10 @@
 from zipfile import ZipFile
 import os
-from database.create_app import app
+from create_app import app
 
 def zip_files(user_id,file_path_list):
-    dir = app.config["SQLITE_DB_DIR"]
-    os.chdir(dir)
     zip_file_name = None
+    flag = True
     try:
         zip_file_name = str(user_id)+"_exported_as_zip.zip"
         
@@ -13,10 +12,12 @@ def zip_files(user_id,file_path_list):
             for file in file_path_list:
                 zip.write(file)
             zip.write('Instructions.pdf')
-    except:
+    except Exception as e:
+        flag = False
         for each in file_path_list:
             os.remove(each)
     finally :
-        for each in file_path_list:
-            os.remove(each)
+        if flag:
+            for each in file_path_list:
+                os.remove(each)
         return zip_file_name
